restructured project into apps/shared/include and apps/{app}/src /build
