+++ 3dd/docker/hguacamole/Dockerfile
@@ -4,13 +4,13 @@ WORKDIR /app
 
 RUN apt-get update && apt-get install -y build-essential meson ninja-build
 
-COPY include/ include/
-COPY src/network/ src/network/
-COPY src/hguacamole/ src/hguacamole/
+COPY apps/shared/include/ apps/shared/include/
+COPY apps/shared/network/ apps/shared/network/
+COPY apps/hguacamole/src apps/hguacamole/src
 
-WORKDIR /app/src/hguacamole
+WORKDIR /app/apps/hguacamole/
 
-RUN meson setup build
+RUN meson setup build src
 RUN meson compile -C build
 
 EXPOSE 5502/udp
