+++ 3dd/docker/ltx_proxy/Dockerfile
@@ -4,13 +4,13 @@ WORKDIR /app
 
 RUN apt-get update && apt-get install -y build-essential meson ninja-build
 
-COPY include/ include/
-COPY src/network/ src/network/
-COPY src/ltx_proxy/ src/ltx_proxy/
+COPY apps/shared/include/ apps/shared/include/
+COPY apps/shared/network/ apps/shared/network/
+COPY apps/ltx_proxy/src apps/ltx_proxy/src
 
-WORKDIR /app/src/ltx_proxy
+WORKDIR /app/apps/ltx_proxy
 
-RUN meson setup build
+RUN meson setup build src/
 RUN meson compile -C build
 
 EXPOSE 5501/udp
